+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Changelog
 [Unreleased]
 ------------
 
+### Added
+
+ - Add CsrfTokenProvider to simplify access to the request token
+
 [3.3.0] (2019-04-09)
 --------------------
 
@@ -100,7 +104,7 @@ Changelog
  - Prevent error than no dca array is loaded (See [contao-leaflet-maps/issues/54](https://github.com/netzmacht/contao-leaflet-maps/issues/54)
  - Added changelog (#9)
 
-[Unreleased]: https://github.com/netzmacht/contao-toolkit/compare/3.3.0...dev-master
+[Unreleased]: https://github.com/netzmacht/contao-toolkit/compare/3.3.0...dev-develop
 [3.3.0]: https://github.com/netzmacht/contao-toolkit/compare/3.2.0...3.3.0
 [3.2.0]: https://github.com/netzmacht/contao-toolkit/compare/3.1.1...3.2.0
 [3.1.1]: https://github.com/netzmacht/contao-toolkit/compare/3.1.0...3.1.1